Исправление замечаний

+++ src/main/java/com/example/api/models/request/GetBooksByAuthorRequest.java
@@ -6,5 +6,5 @@ import lombok.Data;
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @Data
 public class GetBooksByAuthorRequest {
-    private int authorId;
+    private Long authorId;
 }
